requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# HTTP Client & Web Scraping
httpx==0.25.2
beautifulsoup4==4.12.2
lxml==4.9.3

# Task Scheduling
apscheduler==3.10.4

# WebSocket
python-socketio==5.10.0

# Data Validation & Settings
pydantic==2.5.2
pydantic-settings==2.1.0

# Templating
jinja2==3.1.2

# Email
email-validator==2.1.0
aiosmtplib==3.0.1

# Environment Variables
python-dotenv==1.0.0

# AI Integration (Version 2)
openai==1.3.7
anthropic==0.7.8

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
responses==0.24.1

# Development
black==23.12.1
flake8==6.1.0
mypy==1.7.1
